Move element conversion functions to core
