+++ src/poliastro/coordinates.py
@@ -11,7 +11,7 @@ import astropy.units as u
 from astropy.coordinates import get_body_barycentric_posvel, CartesianRepresentation, CartesianDifferential
 
 from poliastro.constants import J2000
-from poliastro.twobody.rv import rv2coe
+from poliastro.core.elements import rv2coe
 from poliastro.util import transform as transform_vector
 from poliastro.twobody.orbit import Orbit
