+++ src/poliastro/twobody/rv.py
@@ -1,66 +1,12 @@
 """Functions to define orbits from position and velocity vectors.
 
 """
-import numpy as np
 from astropy import units as u
 
-from poliastro.core.jit import jit
-from poliastro.core.util import cross, norm
-
 from poliastro.twobody import classical
+from poliastro.core.elements import rv2coe
 
 from ._base import BaseState
-
-
-@jit
-def rv2coe(k, r, v, tol=1e-8):
-    """Converts from vectors to classical orbital elements.
-
-    Parameters
-    ----------
-    k : float
-        Standard gravitational parameter (km^3 / s^2).
-    r : array
-        Position vector (km).
-    v : array
-        Velocity vector (km / s).
-    tol : float, optional
-        Tolerance for eccentricity and inclination checks, default to 1e-8.
-
-    """
-    h = cross(r, v)
-    n = cross([0, 0, 1], h) / norm(h)
-    e = ((v.dot(v) - k / (norm(r))) * r - r.dot(v) * v) / k
-    ecc = norm(e)
-    p = h.dot(h) / k
-    inc = np.arccos(h[2] / norm(h))
-
-    circular = ecc < tol
-    equatorial = abs(inc) < tol
-
-    if equatorial and not circular:
-        raan = 0
-        argp = np.arctan2(e[1], e[0]) % (2 * np.pi)  # Longitude of periapsis
-        nu = (np.arctan2(h.dot(cross(e, r)) / norm(h), r.dot(e)) %
-              (2 * np.pi))
-    elif not equatorial and circular:
-        raan = np.arctan2(n[1], n[0]) % (2 * np.pi)
-        argp = 0
-        # Argument of latitude
-        nu = (np.arctan2(r.dot(cross(h, n)) / norm(h), r.dot(n)) %
-              (2 * np.pi))
-    elif equatorial and circular:
-        raan = 0
-        argp = 0
-        nu = np.arctan2(r[1], r[0]) % (2 * np.pi)  # True longitude
-    else:
-        raan = np.arctan2(n[1], n[0]) % (2 * np.pi)
-        argp = (np.arctan2(e.dot(cross(h, n)) / norm(h), e.dot(n)) %
-                (2 * np.pi))
-        nu = (np.arctan2(r.dot(cross(h, e)) / norm(h), r.dot(e))
-              % (2 * np.pi))
-
-    return p, ecc, inc, raan, argp, nu
 
 
 class RVState(BaseState):
