+++ src/poliastro/twobody/thrust/change_argp.py
@@ -9,7 +9,7 @@ References
 """
 import numpy as np
 
-from poliastro.twobody.rv import rv2coe
+from poliastro.core.elements import rv2coe
 from poliastro.core.util import norm, cross
 from poliastro.core.thrust.change_argp import extra_quantities
